+++ include/anime.h
@@ -163,8 +163,8 @@ int GetAllegienceId(u16);
 void EkrPrepareBanimfx(struct Anim *anim, int);
 // ??? GetSomeAISRelatedIndexMaybeByID(???);
 u32 GetAnimRoundType(int);
-s16 sub_8058A60(int);
-// ??? sub_8058A70(???);
+s16 GetEfxHp(int pos_based_round);
+// ??? GetEfxHpModMaybe(???);
 // ??? IsItemDisplayedInBattle(???);
 // ??? sub_8058AC8(???);
 // ??? sub_8058B08(???);
@@ -189,8 +189,8 @@ void sub_8059D28(void);
 // ??? sub_8059DB8(???);
 // ??? sub_8059E18(???);
 // ??? sub_8059F5C(???);
-void sub_805A07C(struct Anim *anim, int);
-int sub_805A154(void);
+void BanimPrepareSpecificScript(struct Anim *anim, int);
+int sub_805A154(struct Anim *anim);
 int GetAISSubjectId(struct Anim *anim);
 // ??? GetSomeBoolean(???);
 // ??? sub_805A1D0(???);
@@ -215,10 +215,10 @@ void sub_805A940(void*, u16, u16);
 // ??? sub_805A990(???);
 // ??? sub_805A9A4(???);
 void NewEfxAnimeDrvProc(void);
-void sub_805A9E0(void);
+void EndEfxAnimeDrvProc(void);
 // ??? ExecAllAIS(???);
-// ??? sub_805AA00(???);
-// ??? sub_805AA4C(???);
+// ??? NewEkrUnitMainMini(???);
+// ??? EkrUnitMainMiniMain(???);
 void sub_805AA68(void *);
 // ??? sub_805AE40(???);
 // ??? sub_805AE58(???);
@@ -264,7 +264,7 @@ void sub_8071A44(int);
 void sub_8071A8C(void);
 // ??? sub_8071A98(???);
 // ??? sub_8071AA4(???);
-void sub_8071AB0(int, int, int);
+void M4aPlayWithPostionCtrl(int, int, int);
 // ??? sub_8071B6C(???);
 // ??? sub_8072258(???);
 // ??? sub_80723A4(???);
