+++ include/efxbattle.h
@@ -77,22 +77,22 @@ extern struct Unknown_0201FADC gUnknown_0201FADC;
 struct ProcEfxHPBar {
     PROC_HEADER;
 
-    /* 29 */ u8 unk29;
+    /* 29 */ u8 death;
     /* 2A */ u8 _pad_2A[0x2C - 0x2A];
-    /* 2C */ s16 timer;
+    /* 2C */ s16 pos;
     /* 2E */ s16 max;
     /* 30 */ u8 _pad_30[0x48 - 0x30];
-    /* 48 */ int unk48;
-    /* 4C */ int unk4C;
-    /* 50 */ int unk50;
-    /* 54 */ int unk54;
+    /* 48 */ int diff;
+    /* 4C */ int pre;
+    /* 50 */ int post;
+    /* 54 */ int timer;
     /* 58 */ int init;
     /* 5C */ struct Anim *anim5C;
     /* 60 */ struct Anim *anim60;
     /* 64 */ struct Anim *anim64;
 };
 
-void ekrDispUP_Loop(struct ProcEkrDispUP *proc);
+void ekrDispUPMain(struct ProcEkrDispUP *proc);
 
 // ??? NewEkrTogiInitPROC(???);
 // ??? sub_805B104(???);
@@ -821,7 +821,7 @@ void NewEfxDamageMojiEffect(struct Anim *anim, int r1);
 void sub_806E8F0(void);
 void sub_806E904(void);
 void sub_806E920(void);
-// ??? sub_806E93C(???);
+// ??? SetGlbProcefxopCur(???);
 // ??? sub_806E948(???);
 // ??? sub_806E954(???);
 // ??? sub_806E95C(???);
@@ -834,13 +834,13 @@ void sub_806E920(void);
 // ??? sub_806EB2C(???);
 // ??? sub_806EB54(???);
 // ??? nullsub_73(???);
-// ??? sub_806EB7C(???);
+// ??? NewEfxopFire(???);
 // ??? sub_806EB9C(???);
 // ??? sub_806EBBC(???);
 // ??? sub_806EC1C(???);
 // ??? sub_806EC68(???);
 // ??? sub_806ECE8(???);
-// ??? sub_806ED0C(???);
+// ??? NewEfxopThunder(???);
 // ??? sub_806ED2C(???);
 // ??? sub_806ED54(???);
 // ??? sub_806EDB0(???);
