+++ include/ekrdragon.h
@@ -181,7 +181,7 @@ void EkrMyr_WaitForTransform(struct ProcEkrDragon *proc);
 void EkrMyr_TrigerForPrepared(struct ProcEkrDragon *proc);
 void EkrMyr_InBattleIdle(struct ProcEkrDragon *proc);
 void EkrMyr_ReturnToLoli(struct ProcEkrDragon *proc);
-// ??? sub_8070AE4(???);
+void sub_8070AE4(struct ProcEkrDragon *proc);
 void sub_8070B3C(void);
 // ??? sub_8076598(???);
 // ??? CheckEkrWpnDemonLight(???);
@@ -217,7 +217,7 @@ ProcPtr NewEkrDragonBodyAnime(struct Anim *anim);
 // ??? sub_8076F48(???);
 // ??? sub_8076FD0(???);
 // ??? sub_8076FD4(???);
-// ??? EkrDragonBodyAnimeSet54(???);
+void EkrDragonBodyAnimeSet54(struct Anim *anim);
 // ??? EkrDragonBodyAnimeMain(???);
 void NewEkrDragonBodyUnk4(struct Anim *anim, int round_type);
 // ??? EkrDragonBodyUnk4Main(???);
