+++ ldscript.txt
@@ -146,8 +146,9 @@ SECTIONS
         asm/code-banim-08056864.o(.text);
         asm/banim-ekrcheckenable.o(.text);
         asm/banim-ekrmain.o(.text);
-        asm/banim-598CC.o(.text);
+        asm/banim-efx805A07C.o(.text);
         asm/code_battleanim.o(.text);
+        asm/banim-efxop.o(.text);
         src/banim-ekrdragonstatus.o(.text);
         src/banim-ekrdragon-dracozombie.o(.text);
         asm/banim-ekrdragon.o(.text);
@@ -477,6 +478,7 @@ SECTIONS
         . = ALIGN(4); src/banim-efxHPBar.o(.data);
         . = ALIGN(4); data/data_banim_5B93D0.o(.data);
         . = ALIGN(4); src/banim-ekrbattlestarting.o(.data);
+        . = ALIGN(4); data/banim-ekrbattlectrl.o(.data);
         . = ALIGN(4); data/data_banim.o(.data);
         . = ALIGN(4); src/banim-ekrdragon-dracozombie.o(.data);
         . = ALIGN(4); data/data-ekrdragon.o(.data);
