+++ src/banim-efxHPBar.c
@@ -10,236 +10,9 @@
 #include "ekrdragon.h"
 #include "constants/items.h"
 
+/* Banim hp-bar change effect */
+
 extern struct Anim *gAnims[4];
-
-void EfxClearScreenFx(void)
-{
-    gLCDControlBuffer.dispcnt.mode = 0;
-    SetDispEnable(1, 1, 1, 1, 1);
-    gLCDControlBuffer.dispcnt.obj1dMap = 0;
-
-    BG_SetPosition(BG_0, 0, 0);
-    BG_SetPosition(BG_1, 0, 0);
-    BG_SetPosition(BG_2, 0, 0);
-
-    SetBackgroundTileDataOffset(BG_0, 0);
-    SetBackgroundTileDataOffset(BG_1, 0);
-    SetBackgroundTileDataOffset(BG_2, 0);
-    SetBackgroundTileDataOffset(BG_3, 0x8000);
-
-    SetBackgroundMapDataOffset(BG_0, 0x6000);
-    SetBackgroundMapDataOffset(BG_1, 0x6800);
-    SetBackgroundMapDataOffset(BG_2, 0x7000);
-    SetBackgroundMapDataOffset(BG_3, 0x7800);
-
-    gLCDControlBuffer.bg0cnt.priority = 0;
-    gLCDControlBuffer.bg1cnt.priority = 1;
-    gLCDControlBuffer.bg2cnt.priority = 2;
-    gLCDControlBuffer.bg3cnt.priority = 3;
-
-    CpuFastFill16(0, gBG0TilemapBuffer, 0x800);
-    CpuFastFill16(0, gBG1TilemapBuffer, 0x800);
-    CpuFastFill16(0, gBG2TilemapBuffer, 0x800);
-
-    if (GetBattleAnimArenaFlag() == false)
-        sub_8051E00();
-    else
-        CpuFastFill16(0, gBG2TilemapBuffer, 0x800);
-
-    EfxPrepareScreenFx();
-    EnablePaletteSync();
-
-    BG_EnableSyncByMask(BG0_SYNC_BIT);
-    BG_EnableSyncByMask(BG1_SYNC_BIT);
-    BG_EnableSyncByMask(BG2_SYNC_BIT);
-    SetDefaultColorEffects();
-}
-
-void sub_8051E00(void)
-{
-    struct Unknown_0201FADC *unk0201FADC = &gUnknown_0201FADC;
-    BattleAnimTerrain *terrain1 = &battle_terrain_table[gBanimTerrainIndexMaybe[0]];
-    BattleAnimTerrain *terrain2 = &battle_terrain_table[gBanimTerrainIndexMaybe[1]];
-
-    switch (gEkrDistanceType) {
-    case EKR_DISTANCE_CLOSE:
-    case EKR_DISTANCE_PROMOTION:
-        gUnknown_0200003C[0] = &gUnknown_020145C8[0];
-        gUnknown_0200003C[1] = &gUnknown_020145C8[0x1000];
-        break;
-
-    case EKR_DISTANCE_FAR:
-    case EKR_DISTANCE_FARFAR:
-    case EKR_DISTANCE_3:
-        gUnknown_0200003C[0] = &gUnknown_020145C8[0x800];
-        gUnknown_0200003C[1] = &gUnknown_020145C8[0x1800];
-        break;
-
-    }
-
-    switch (gPlaySt.chapterWeatherId) {
-        // dummy, both cases do the same thing
-    case WEATHER_SNOW:
-        gBanimTerrainPaletteMaybe[0] = terrain1->palette;
-        gBanimTerrainPaletteMaybe[1] = terrain2->palette;
-        break;
-
-    default:
-        gBanimTerrainPaletteMaybe[0] = terrain1->palette;
-        gBanimTerrainPaletteMaybe[1] = terrain2->palette;
-        break;
-    }
-
-    gUnknown_02000044[0] = gUnknown_085B9D6C[gEkrDistanceType * 2];
-    gUnknown_02000044[1] = gUnknown_085B9D6C[gEkrDistanceType * 2 + 1];
-
-    unk0201FADC->things[0] = gBanimTerrainIndexMaybe[0];
-    unk0201FADC->things[1] = 4;
-    unk0201FADC->things[2] = 640;
-    unk0201FADC->things[3] = gBanimTerrainIndexMaybe[1];
-    unk0201FADC->things[4] = 5;
-    unk0201FADC->things[5] = 640;
-    unk0201FADC->things[6] = gEkrDistanceType;
-    unk0201FADC->things[7] = 2;
-    unk0201FADC->u1c = 0;
-    unk0201FADC->u20 = &gUnknown_020145C8[0];
-    unk0201FADC->_pad_10 = gEkrSnowWeather;
-    sub_805AA68(unk0201FADC);
-}
-
-void EfxPrepareScreenFx(void)
-{
-    const char *str;
-
-    CopyToPaletteBuffer(Pal_UIFont, 0x40, 0x20);
-    CopyToPaletteBuffer(Pal_UIFont, 0x60, 0x20);
-    Font_InitForUI(&gSomeFontStruct, (void *)0x6001880, 0xC4, 2);
-    Font_SetSomeSpecialDrawingRoutine();
-    LZ77UnCompVram(gUnknown_08801C14, (void *)0x6001000);
-
-    /* left unit name */
-    if (gEkrPairSideVaild[EKR_BATTLE_LEFT] == false)
-        str = gNopStr;
-    else
-        str = GetStringFromIndex(gpEkrBattleUnitLeft->unit.pCharacterData->nameTextId);
-
-    Text_Init(&gTextEkrlvupMsg[0], 7);
-    Text_SetXCursor(&gTextEkrlvupMsg[0], GetStringTextCenteredPos(0x38, str));
-    LZ77UnCompVram(Img_EfxLeftNameBox, (void *)0x6001880);
-    Text_AppendString(&gTextEkrlvupMsg[0], str);
-
-    /* left unit item */
-    if (gEkrPairSideVaild[EKR_BATTLE_LEFT] == false)
-        str = gNopStr;
-    else
-        str = GetItemName(gpEkrBattleUnitLeft->weaponBefore);
-
-    Text_Init(&gTextEkrlvupMsg[2], 8);
-    Text_SetXCursor(&gTextEkrlvupMsg[2], GetStringTextCenteredPos(0x40, str));
-    LZ77UnCompVram(Img_EfxLeftItemBox, (void *)0x6001A40);
-    Text_AppendString(&gTextEkrlvupMsg[2], str);
-
-    /* right unit name */
-    if (gEkrPairSideVaild[EKR_BATTLE_RIGHT] == false)
-        str = gNopStr;
-    else
-        str = GetStringFromIndex(gpEkrBattleUnitRight->unit.pCharacterData->nameTextId);
-
-    Text_Init(&gTextEkrlvupMsg[3], 7);
-    Text_SetXCursor(&gTextEkrlvupMsg[3], GetStringTextCenteredPos(0x38, str));
-    LZ77UnCompVram(Img_EfxRightNameBox, (void *)0x6001C40);
-    Text_AppendString(&gTextEkrlvupMsg[3], str);
-
-    /* right unit item */
-    if (gEkrPairSideVaild[EKR_BATTLE_RIGHT] == false)
-        str = gNopStr;
-    else
-        str = GetItemName(gpEkrBattleUnitRight->weaponBefore);
-
-    Text_Init(&gTextEkrlvupMsg[1], 8);
-    Text_SetXCursor(&gTextEkrlvupMsg[1], GetStringTextCenteredPos(0x3E, str));
-    LZ77UnCompVram(Img_EfxRightItemBox, (void *)0x6001E00);
-    Text_AppendString(&gTextEkrlvupMsg[1], str);
-
-    BG_Fill(gBG0TilemapBuffer, 0x80);
-    sub_8070E94(gUnknown_08802508, gBG0TilemapBuffer + 0x1E, 2, 20, -1, -1);
-    sub_8070D04(gBG0TilemapBuffer + 0x1F, 1, 20, 2, 128);
-    sub_8070D04(gBG0TilemapBuffer + 0x1E, 1, 20, 3, 128);
-    BG_EnableSyncByMask(BG0_SYNC_BIT);
-
-    CpuFastCopy(&PAL_BUF_COLOR(gUnknown_08802558, gBanimSomeObjPalIndex[0], 0), PAL_BG(0x2), 0x20);
-    CpuFastCopy(&PAL_BUF_COLOR(gUnknown_08802558, gBanimSomeObjPalIndex[1], 0), PAL_BG(0x3), 0x20);
-    EnablePaletteSync();
-
-    gEkrBg0QuakeVec.x = 0;
-    gEkrBg0QuakeVec.y = 0;
-    BG_SetPosition(BG_0, 0, 0);
-}
-
-int GetEkrSomePosMaybe(void)
-{
-    int quote1, quote2;
-
-    switch (gEkrDistanceType) {
-    case 1:
-        return gEkrInitialHitSide;
-
-    case 0:
-    case 3:
-    case 4:
-        return EKR_BATTLE_RIGHT;
-
-    case 2:
-    default:
-        quote2 = false;
-        quote1 = false;
-
-        if (gUnknown_0203E0FC == 0) {
-            quote1 = ShouldCallBattleQuote(gEkrPids[EKR_BATTLE_LEFT],  gEkrPids[EKR_BATTLE_RIGHT]);
-            quote2 = ShouldCallBattleQuote(gEkrPids[EKR_BATTLE_RIGHT], gEkrPids[EKR_BATTLE_LEFT]);
-        }
-
-        if (quote1 == true)
-            return EKR_BATTLE_LEFT;
-        else if (quote2 == true)
-            return EKR_BATTLE_RIGHT;
-        else
-            return gEkrInitialHitSide;
-    }
-}
-
-void sub_8052214(int a, int b)
-{
-    gEkrBg2QuakeVec.x = a;
-    gEkrBg2QuakeVec.y = b;
-}
-
-void EkrEfxStatusClear(void)
-{
-    gEkrHPBarCount = 0;
-    gUnknown_0201772C = 0;
-    gUnknown_02017730 = 0;
-    gEkrDeadEventFlag = 0;
-    gUnknown_0201773C = 0;
-    gUnknown_02017740 = 0;
-    gUnknown_02017748 = 0;
-    gUnknown_0201774C = 0;
-    gUnknown_02017750 = 0;
-    gUnknown_02017754 = 0;
-    gUnknown_02017758 = 0;
-    gUnknown_0201775C = 0;
-    sub_8052214(0, 0);
-    gUnknown_02017764[0] = 0;
-    gUnknown_02017764[1] = 0;
-    gUnknown_02017768[0] = 0;
-    gUnknown_02017768[1] = 0;
-    gEkrHitEfxBool[0] = 0;
-    gEkrHitEfxBool[1] = 0;
-    gUnknown_0201776C[0] = 0;
-    gUnknown_0201776C[1] = 0;
-    gUnknown_02017778 = 0;
-    gpProcEfxHPBarColorChange = 0;
-}
 
 int sub_80522CC(void)
 {
@@ -254,6 +27,7 @@ short EkrEfxIsUnitHittedNow(int pos)
     return gEkrHitEfxBool[pos];
 }
 
+/* During dmage, cause hp bar change value effect */
 void NewEfxHPBar(struct Anim *anim)
 {
     int val1;
@@ -269,53 +43,53 @@ void NewEfxHPBar(struct Anim *anim)
     proc = Proc_Start(ProcScr_efxHPBar, PROC_TREE_3);
     proc->anim64 = anim;
 
-    if (GetAISSubjectId(anim) == 0) {
-        proc->anim5C = gAnims[2];
-        proc->anim60 = gAnims[0];
+    if (GetAISSubjectId(anim) == EKR_BATTLE_LEFT) {
+        proc->anim5C = gAnims[EKR_BATTLE_RIGHT * 2];
+        proc->anim60 = gAnims[EKR_BATTLE_LEFT * 2];
     } else {
-        proc->anim5C = gAnims[0];
-        proc->anim60 = gAnims[2];
+        proc->anim5C = gAnims[EKR_BATTLE_LEFT * 2];
+        proc->anim60 = gAnims[EKR_BATTLE_RIGHT * 2];
     }
 
-    val1 = gUnknown_0203E152[GetAISSubjectId(proc->anim60)];
+    val1 = gEfxPairHpBufOffset[GetAISSubjectId(proc->anim60)];
     val2 = val1 + 1;
 
-    proc->unk4C = sub_8058A60(val1 * 2 + GetAISSubjectId(proc->anim60));
-    proc->unk50 = sub_8058A60(val2 * 2 + GetAISSubjectId(proc->anim60));
+    proc->pre = GetEfxHp(val1 * 2 + GetAISSubjectId(proc->anim60));
+    proc->post = GetEfxHp(val2 * 2 + GetAISSubjectId(proc->anim60));
 
-    if (proc->unk4C > proc->unk50)
-        proc->unk48 = -1;
+    if (proc->pre > proc->post)
+        proc->diff = -1;
     else
-        proc->unk48 = 1;
+        proc->diff = 1;
 
+    proc->pos = 0;
+    proc->max = proc->pre;
     proc->timer = 0;
-    proc->max = proc->unk4C;
-    proc->unk54 = 0;
     proc->init = false;
     gEkrHitEfxBool[GetAISSubjectId(proc->anim60)] = 1;
 }
 
-void efxHPBar_80523EC(struct ProcEfxHPBar *proc)
+void EfxHp_BarDeclineWithDeathJudge(struct ProcEfxHPBar *proc)
 {
     struct Anim *anim1 = gAnims[GetAISSubjectId(proc->anim60) * 2];
     struct Anim *anim2 = gAnims[GetAISSubjectId(proc->anim60) * 2 + 1];
 
     if (proc->init == false) {
-        if (++proc->timer == 2) {
-            proc->timer = 0;
-            proc->max += proc->unk48;
-            gEkrPairHpInitial[GetAISSubjectId(proc->anim60)] += proc->unk48;
+        if (++proc->pos == 2) {
+            proc->pos = 0;
+            proc->max += proc->diff;
+            gEkrPairHpInitial[GetAISSubjectId(proc->anim60)] += proc->diff;
 
-            if (proc->max == proc->unk50)
+            if (proc->max == proc->post)
                 proc->init = true;
         }
     }
 
-    if (proc->unk54 == 0x1E && proc->init == true) {
-        gUnknown_0203E152[GetAISSubjectId(proc->anim60)]++;
+    if (proc->timer == 0x1E && proc->init == true) {
+        gEfxPairHpBufOffset[GetAISSubjectId(proc->anim60)]++;
         gEkrHitEfxBool[GetAISSubjectId(proc->anim60)] = 0;
     
-        if (proc->unk50 == 0) {
+        if (proc->post == 0) {
             int ret;
             if (GetBanimLinkArenaFlag() == true)
                 ret = 0;
@@ -330,13 +104,13 @@ void efxHPBar_80523EC(struct ProcEfxHPBar *proc)
                 gEkrPairSideVaild[GetAISSubjectId(proc->anim60)] = false;
             }
         }
-    
+
         Proc_Break(proc);
         return;
     }
 
-    if (++proc->unk54 >= 0x1Eu)
-        proc->unk54 = 0x1E;
+    if (++proc->timer >= 0x1Eu)
+        proc->timer = 0x1E;
 }
 
 void efxHPBarMain(struct ProcEfxHPBar *proc)
@@ -344,23 +118,23 @@ void efxHPBarMain(struct ProcEfxHPBar *proc)
     struct Anim *anim;
 
     if (gUnknown_0201774C == 0 && gUnknown_0201772C == 0) {
-        proc->timer = 0;
+        proc->pos = 0;
         proc->max = 1;
         anim = GetCoreAIStruct(proc->anim64);
 
         if (sub_805A21C(sub_805A2F0(anim)) == 1) {
             switch (gEkrDistanceType) {
-            case 0:
-            case 1:
-            case 3:
-            case 4:
-                proc->max = 0x10;
-                sub_80533D0(GetCoreAIStruct(anim), -1);
+            case EKR_DISTANCE_CLOSE:
+            case EKR_DISTANCE_FAR:
+            case EKR_DISTANCE_3:
+            case EKR_DISTANCE_PROMOTION:
+                proc->max = 16;
+                NewEfxFarAttackWithDistance(GetCoreAIStruct(anim), -1);
                 break;
 
-            case 2:
-                proc->max = 0x14;
-                sub_80533D0(GetCoreAIStruct(anim), -1);
+            case EKR_DISTANCE_FARFAR:
+                proc->max = 20;
+                NewEfxFarAttackWithDistance(GetCoreAIStruct(anim), -1);
                 break;
             }
         }
@@ -369,12 +143,12 @@ void efxHPBarMain(struct ProcEfxHPBar *proc)
     }
 }
 
-void efxHPBarMainEnd(struct ProcEfxHPBar *proc)
+void efxHPBarWaitForFarFarCamMoveMaybe(struct ProcEfxHPBar *proc)
 {
-    if (++proc->timer == (proc->max - 4)) {
+    if (++proc->pos == (proc->max - 4)) {
         GetCoreAIStruct(proc->anim64);
         BG_EnableSyncByMask(BG2_SYNC_BIT);
-    } else if (proc->timer == proc->max){
+    } else if (proc->pos == proc->max){
         gEkrHPBarCount--;
         Proc_Break(proc);
     }
@@ -394,29 +168,29 @@ void NewEfxHPBarResire(struct Anim *anim)
     proc = Proc_Start(ProcScr_efxHPBarResire, PROC_TREE_3);
     proc->anim64 = GetCoreAIStruct(anim);
 
-    if (GetAISSubjectId(anim) == 0) {
-        proc->anim5C = gAnims[2];
-        proc->anim60 = gAnims[0];
+    if (GetAISSubjectId(anim) == EKR_BATTLE_LEFT) {
+        proc->anim5C = gAnims[EKR_BATTLE_RIGHT * 2];
+        proc->anim60 = gAnims[EKR_BATTLE_LEFT * 2];
     } else {
-        proc->anim5C = gAnims[0];
-        proc->anim60 = gAnims[2];
+        proc->anim5C = gAnims[EKR_BATTLE_LEFT * 2];
+        proc->anim60 = gAnims[EKR_BATTLE_RIGHT * 2];
     }
 
-    val1 = gUnknown_0203E152[GetAISSubjectId(proc->anim60)];
+    val1 = gEfxPairHpBufOffset[GetAISSubjectId(proc->anim60)];
     val2 = val1 + 1;
 
-    proc->unk4C = sub_8058A60(val1 * 2 + GetAISSubjectId(proc->anim60));
-    proc->unk50 = sub_8058A60(val2 * 2 + GetAISSubjectId(proc->anim60));
+    proc->pre = GetEfxHp(val1 * 2 + GetAISSubjectId(proc->anim60));
+    proc->post = GetEfxHp(val2 * 2 + GetAISSubjectId(proc->anim60));
 
-    if (proc->unk4C > proc->unk50)
-        proc->unk48 = -1;
+    if (proc->pre > proc->post)
+        proc->diff = -1;
     else
-        proc->unk48 = 1;
+        proc->diff = 1;
 
-    proc->unk29 = 0;
+    proc->death = false;
+    proc->pos = 0;
+    proc->max = proc->pre;
     proc->timer = 0;
-    proc->max = proc->unk4C;
-    proc->unk54 = 0;
     proc->init = false;
     gUnknown_02017750 = 0;
     gEkrHitEfxBool[GetAISSubjectId(proc->anim60)] = 1;
@@ -428,33 +202,33 @@ void EfxHPBarResire_80526C8(struct ProcEfxHPBar *proc)
     GetAISSubjectId(proc->anim60);
 
     if (proc->init == false) {
-        if (++proc->timer == 2) {
-            proc->timer = 0;
-            proc->max += proc->unk48;
-            gEkrPairHpInitial[GetAISSubjectId(proc->anim60)] += proc->unk48;
+        if (++proc->pos == 2) {
+            proc->pos = 0;
+            proc->max += proc->diff;
+            gEkrPairHpInitial[GetAISSubjectId(proc->anim60)] += proc->diff;
 
-            if (proc->max == proc->unk50)
+            if (proc->max == proc->post)
                 proc->init = true;
         }
     }
 
-    if (proc->unk54 == 0x54 && proc->init == true) {
-        gUnknown_0203E152[GetAISSubjectId(proc->anim60)]++;
+    if (proc->timer == 0x54 && proc->init == true) {
+        gEfxPairHpBufOffset[GetAISSubjectId(proc->anim60)]++;
         gEkrHitEfxBool[GetAISSubjectId(proc->anim60)] = 0;
 
-        if (proc->unk50 == 0)
-            proc->unk29 = 1;
+        if (proc->post == 0)
+            proc->death = true;
 
-        proc->timer = 0;
-        proc->max = 0xA;
+        proc->pos = 0;
+        proc->max = 10;
         gUnknown_02017750 = 1;
 
         Proc_Break(proc);
         return;
     }
 
-    if (++proc->unk54 >= 0x54u)
-        proc->unk54 = 0x54;
+    if (++proc->timer >= 0x54u)
+        proc->timer = 0x54;
 }
 
 void EfxHPBarResire_8052788(struct ProcEfxHPBar *proc)
@@ -462,27 +236,27 @@ void EfxHPBarResire_8052788(struct ProcEfxHPBar *proc)
     int val1;
     s16 val2;
 
-    if (++proc->timer <= proc->max)
+    if (++proc->pos <= proc->max)
         return;
 
-    val1 = gUnknown_0203E152[GetAISSubjectId(proc->anim5C)];
+    val1 = gEfxPairHpBufOffset[GetAISSubjectId(proc->anim5C)];
     val2 = val1 + 1;
 
-    proc->unk4C = sub_8058A60(val1 * 2 + GetAISSubjectId(proc->anim5C));
-    proc->unk50 = sub_8058A60(val2 * 2 + GetAISSubjectId(proc->anim5C));
+    proc->pre = GetEfxHp(val1 * 2 + GetAISSubjectId(proc->anim5C));
+    proc->post = GetEfxHp(val2 * 2 + GetAISSubjectId(proc->anim5C));
 
+    proc->pos = 0;
+    proc->max = proc->pre;
     proc->timer = 0;
-    proc->max = proc->unk4C;
-    proc->unk54 = 0;
     proc->init = false;
 
-    if (proc->unk4C == proc->unk50)
+    if (proc->pre == proc->post)
         proc->init = true;
 
-    if (proc->unk4C > proc->unk50)
-        proc->unk48 = -1;
+    if (proc->pre > proc->post)
+        proc->diff = -1;
     else
-        proc->unk48 = 1;
+        proc->diff = 1;
 
     Proc_Break(proc);
     gEkrHitEfxBool[GetAISSubjectId(proc->anim5C)] = 2;
@@ -498,23 +272,23 @@ void EfxHPBarResire_805282C(struct ProcEfxHPBar *proc)
     anim4 = gAnims[GetAISSubjectId(proc->anim60) * 2 + 1];
 
     if (proc->init == false) {
-        if (++proc->timer == 4) {
-            proc->timer = 0;
-            proc->max += proc->unk48;
-            gEkrPairHpInitial[GetAISSubjectId(proc->anim5C)] += proc->unk48;
+        if (++proc->pos == 4) {
+            proc->pos = 0;
+            proc->max += proc->diff;
+            gEkrPairHpInitial[GetAISSubjectId(proc->anim5C)] += proc->diff;
             SomePlaySound_8071990(0x75, 0x100);
-            sub_8071AB0(0x75, anim1->xPosition, 1);
+            M4aPlayWithPostionCtrl(0x75, anim1->xPosition, 1);
 
-            if (proc->max == proc->unk50)
+            if (proc->max == proc->post)
                 proc->init = true;
         }
     }
 
-    if (proc->unk54 == 0x1E && proc->init == true) {
-        gUnknown_0203E152[GetAISSubjectId(proc->anim5C)]++;
+    if (proc->timer == 30 && proc->init == true) {
+        gEfxPairHpBufOffset[GetAISSubjectId(proc->anim5C)]++;
         gEkrHitEfxBool[GetAISSubjectId(proc->anim5C)] = 0;
     
-        if (proc->unk29 == 1) {
+        if (proc->death == true) {
             int ret;
             if (GetBanimLinkArenaFlag() == true)
                 ret = 0;
@@ -534,8 +308,8 @@ void EfxHPBarResire_805282C(struct ProcEfxHPBar *proc)
         return;
     }
 
-    if (++proc->unk54 >= 0x1Eu)
-        proc->unk54 = 0x1E;
+    if (++proc->timer >= 0x1Eu)
+        proc->timer = 0x1E;
 }
 
 void NewEfxAvoid(struct Anim *anim)
@@ -550,7 +324,7 @@ void NewEfxAvoid(struct Anim *anim)
     gEkrHPBarCount = 1;
 
     proc = Proc_Start(ProcScr_EfxAvoid, PROC_TREE_3);
-    proc->timer = 0;
+    proc->pos = 0;
 
     if (GetAISSubjectId(anim) == 0) {
         proc->anim5C = gAnims[2];
@@ -562,14 +336,14 @@ void NewEfxAvoid(struct Anim *anim)
 
     NewEfxDamageMojiEffect(proc->anim60, 1);
     proc->anim64 = anim;
-    proc->unk29 = 0;
+    proc->death = false;
     SomePlaySound_8071990(0xD7, 0x100);
-    sub_8071AB0(0xD7, anim->xPosition, 1);
+    M4aPlayWithPostionCtrl(0xD7, anim->xPosition, 1);
 }
 
 void EfxAvoidMain(struct ProcEfxHPBar *proc)
 {
-    if (++proc->timer == 0x1E)
+    if (++proc->pos == 0x1E)
         Proc_Break(proc);
 }
 
@@ -595,24 +369,24 @@ void NewEfxHPBarLive(struct Anim *anim)
         proc->anim60 = gAnims[2];
     }
 
-    val1 = gUnknown_0203E152[GetAISSubjectId(proc->anim60)];
+    val1 = gEfxPairHpBufOffset[GetAISSubjectId(proc->anim60)];
     val2 = val1 + 1;
 
-    proc->unk4C = sub_8058A60(val1 * 2 + GetAISSubjectId(proc->anim60));
-    proc->unk50 = sub_8058A60(val2 * 2 + GetAISSubjectId(proc->anim60));
-
-    proc->unk54 = 0;
-    proc->init = false;
-    
-    if (proc->unk4C == proc->unk50)
-        proc->init = true;
-    else if (proc->unk4C > proc->unk50)
-        proc->unk48 = -1;
-    else
-        proc->unk48 = 1;
+    proc->pre = GetEfxHp(val1 * 2 + GetAISSubjectId(proc->anim60));
+    proc->post = GetEfxHp(val2 * 2 + GetAISSubjectId(proc->anim60));
 
     proc->timer = 0;
-    proc->max = proc->unk4C;
+    proc->init = false;
+    
+    if (proc->pre == proc->post)
+        proc->init = true;
+    else if (proc->pre > proc->post)
+        proc->diff = -1;
+    else
+        proc->diff = 1;
+
+    proc->pos = 0;
+    proc->max = proc->pre;
     proc->anim64 = anim;
     gEkrHitEfxBool[GetAISSubjectId(proc->anim5C)] = 2;
 }
@@ -622,31 +396,31 @@ void EfxHPBarLiveMain(struct ProcEfxHPBar *proc)
     struct Anim *anim = proc->anim60;
 
     if (proc->init == false) {
-        if (++proc->timer == 4) {
-            proc->timer = 0;
-            proc->max += proc->unk48;
-            gEkrPairHpInitial[GetAISSubjectId(anim)] += proc->unk48;
+        if (++proc->pos == 4) {
+            proc->pos = 0;
+            proc->max += proc->diff;
+            gEkrPairHpInitial[GetAISSubjectId(anim)] += proc->diff;
 
             SomePlaySound_8071990(0x75, 0x100);
-            sub_8071AB0(0x75, anim->xPosition, 1);
+            M4aPlayWithPostionCtrl(0x75, anim->xPosition, 1);
 
-            if (proc->max == proc->unk50)
+            if (proc->max == proc->post)
                 proc->init = true;
         }
     }
 
-    if (proc->unk54 == 0x1E && proc->init == true) {
-        gUnknown_0203E152[GetAISSubjectId(anim)]++;
+    if (proc->timer == 0x1E && proc->init == true) {
+        gEfxPairHpBufOffset[GetAISSubjectId(anim)]++;
         gEkrHitEfxBool[GetAISSubjectId(anim)] = 0;
         Proc_Break(proc);
         return;
     }
 
-    if (++proc->unk54 >= 0x1Eu)
-        proc->unk54 = 0x1E;
+    if (++proc->timer >= 0x1Eu)
+        proc->timer = 0x1E;
 }
 
-void NewEfxNoDmage(struct Anim *anim1, struct Anim *anim2, int arg)
+void NewEfxNoDmage(struct Anim *anim1, struct Anim *anim2, int death)
 {
     struct BattleUnit *bu;
     struct ProcEfxHPBar *proc;
@@ -665,8 +439,8 @@ void NewEfxNoDmage(struct Anim *anim1, struct Anim *anim2, int arg)
         proc = Proc_Start(ProcScr_efxNoDamage, PROC_TREE_3);
         proc->anim5C = anim1;
         proc->anim60 = anim2;
-        proc->timer = 0;
-        proc->unk29 = arg;
+        proc->pos = 0;
+        proc->death = death;
         proc->anim64 = anim1;
         NewEfxDamageMojiEffect(proc->anim5C, 0);
         NewEfxNoDamageYure(proc->anim5C, proc->anim60);
@@ -678,13 +452,13 @@ void EfxNoDamageMain(struct ProcEfxHPBar *proc)
     s32 val;
     struct Anim *anim = GetCoreAIStruct(proc->anim5C);
 
-    ++proc->timer;
-    val = proc->timer;
+    ++proc->pos;
+    val = proc->pos;
     if (val == 8) {
-        gUnknown_0203E152[GetAISSubjectId(proc->anim5C)]++;
+        gEfxPairHpBufOffset[GetAISSubjectId(proc->anim5C)]++;
     
-        if (proc->unk29 == 1)
-            gUnknown_0203E152[GetAISSubjectId(anim)]++;
+        if (proc->death == true)
+            gEfxPairHpBufOffset[GetAISSubjectId(anim)]++;
     
         Proc_Break(proc);
     }
@@ -695,7 +469,7 @@ void NewEfxNoDamageYure(struct Anim *anim1, struct Anim *anim2)
     struct ProcEfxHPBar *proc = Proc_Start(ProcScr_efxNoDamageYure, PROC_TREE_3);
     proc->anim5C = anim1;
     proc->anim60 = anim2;
-    proc->timer = 0;
+    proc->pos = 0;
     proc->max = 0;
 }
 
@@ -705,7 +479,7 @@ void EfxNoDamageYureMain(struct ProcEfxHPBar *proc)
     struct Anim *anim1 = proc->anim5C;
     struct Anim *anim2 = proc->anim60;
 
-    if (gUnknown_080DA438[proc->timer] == -1) {
+    if (gEfxNoDmgBgShakeOff[proc->pos] == -1) {
         if (GetBanimDragonStatusType() == EKRDRGON_TYPE_NORMAL) {
             anim1->xPosition = gEkrXPosBase[GetAISSubjectId(proc->anim5C)] - gEkrBgXOffset;
             anim2->xPosition = gEkrXPosBase[GetAISSubjectId(proc->anim60)] - gEkrBgXOffset;
@@ -714,15 +488,15 @@ void EfxNoDamageYureMain(struct ProcEfxHPBar *proc)
         Proc_Break(proc);
     } else {
         if (GetAISSubjectId(anim1) == 1)
-            val1 = -gUnknown_080DA438[proc->timer];
+            val1 = -gEfxNoDmgBgShakeOff[proc->pos];
         else
-            val1 = gUnknown_080DA438[proc->timer];
+            val1 = gEfxNoDmgBgShakeOff[proc->pos];
 
         if (GetBanimDragonStatusType() == EKRDRGON_TYPE_NORMAL) {
             anim1->xPosition = gEkrXPosBase[GetAISSubjectId(proc->anim5C)] - (s32)gEkrBgXOffset + val1;
             anim2->xPosition = gEkrXPosBase[GetAISSubjectId(proc->anim60)] - (s32)gEkrBgXOffset + val1;
         }
-        proc->timer++;
+        proc->pos++;
     }
 }
 
@@ -733,14 +507,14 @@ void NewEfxStatusCHG(struct Anim *anim)
     if (gEkrHPBarCount == 0) {
         gEkrHPBarCount = 1;
         proc = Proc_Start(ProcScr_efxStatusCHG, PROC_TREE_3);
-        proc->timer = 0;
+        proc->pos = 0;
         proc->anim64 = anim;
     }
 }
 
 void EfxStatusCHGMain(struct ProcEfxHPBar *proc)
 {
-    if (++proc->timer == 0x11)
+    if (++proc->pos == 0x11)
         Proc_Break(proc);
 }
 
@@ -750,9 +524,9 @@ void EfxStatusCHGMain(struct ProcEfxHPBar *proc)
 
 CONST_DATA struct ProcCmd ProcScr_efxHPBar[] = {
     PROC_NAME("efxHPBar"),
-    PROC_REPEAT(efxHPBar_80523EC),
+    PROC_REPEAT(EfxHp_BarDeclineWithDeathJudge),
     PROC_REPEAT(efxHPBarMain),
-    PROC_REPEAT(efxHPBarMainEnd),
+    PROC_REPEAT(efxHPBarWaitForFarFarCamMoveMaybe),
     PROC_END
 };
 
@@ -762,7 +536,7 @@ CONST_DATA struct ProcCmd ProcScr_efxHPBarResire[] = {
     PROC_REPEAT(EfxHPBarResire_8052788),
     PROC_REPEAT(EfxHPBarResire_805282C),
     PROC_REPEAT(efxHPBarMain),
-    PROC_REPEAT(efxHPBarMainEnd),
+    PROC_REPEAT(efxHPBarWaitForFarFarCamMoveMaybe),
     PROC_END
 };
 
@@ -770,7 +544,7 @@ CONST_DATA struct ProcCmd ProcScr_EfxAvoid[] = {
     PROC_NAME("efxAvoid"),
     PROC_REPEAT(EfxAvoidMain),
     PROC_REPEAT(efxHPBarMain),
-    PROC_REPEAT(efxHPBarMainEnd),
+    PROC_REPEAT(efxHPBarWaitForFarFarCamMoveMaybe),
     PROC_END
 };
 
@@ -778,7 +552,7 @@ CONST_DATA struct ProcCmd ProcScr_efxHPBarLive[] = {
     PROC_NAME("efxHPBarLive"),
     PROC_REPEAT(EfxHPBarLiveMain),
     PROC_REPEAT(efxHPBarMain),
-    PROC_REPEAT(efxHPBarMainEnd),
+    PROC_REPEAT(efxHPBarWaitForFarFarCamMoveMaybe),
     PROC_END
 };
 
@@ -786,7 +560,7 @@ CONST_DATA struct ProcCmd ProcScr_efxNoDamage[] = {
     PROC_NAME("efxNoDamage"),
     PROC_REPEAT(EfxNoDamageMain),
     PROC_REPEAT(efxHPBarMain),
-    PROC_REPEAT(efxHPBarMainEnd),
+    PROC_REPEAT(efxHPBarWaitForFarFarCamMoveMaybe),
     PROC_END
 };
 
@@ -796,7 +570,7 @@ CONST_DATA struct ProcCmd ProcScr_efxNoDamageYure[] = {
     PROC_END
 };
 
-const s16 gUnknown_080DA438[] = {
+const s16 gEfxNoDmgBgShakeOff[] = {
     0, 1, 2, 3, 2, 1, 0,
     0, 0, 0, 0, 0, 0, -1
 };
@@ -806,6 +580,6 @@ CONST_DATA struct ProcCmd ProcScr_efxStatusCHG[] = {
     PROC_MARK(0xA),
     PROC_REPEAT(EfxStatusCHGMain),
     PROC_REPEAT(efxHPBarMain),
-    PROC_REPEAT(efxHPBarMainEnd),
+    PROC_REPEAT(efxHPBarWaitForFarFarCamMoveMaybe),
     PROC_END
 };
