+++ src/banim-ekrlvupfan.c
@@ -17,7 +17,7 @@ void EkrLvupFanMain(struct ProcEkrLvupFan *proc)
     int timer = ++proc->timer;
     if (timer == 0x10) {
         SomePlaySound_8071990(0x5B, 0x100);
-        sub_8071AB0(0x5B, 0x78, 0);
+        M4aPlayWithPostionCtrl(0x5B, 0x78, 0);
     } else if (timer == 0x74) {
         Sound_SetSEVolume(0x100);
         Proc_Break(proc);
