+++ sym_ewram.txt
@@ -13,8 +13,8 @@
 . = 0x000044; gUnknown_02000044 = .;
 . = 0x00004C; gBanimTerrainPaletteMaybe = .;
 . = 0x000054; gUnknown_02000054 = .;
-. = 0x00005C; gUnknown_0200005C = .;
-. = 0x000060; gUnknown_02000060 = .;
+. = 0x00005C; gpBanimModesLeft = .;
+. = 0x000060; gpBanimModesRight = .;
 . = 0x000064; gpProcEkrBattle = .;
 . = 0x000068; gpProcEkrGauge = .;
 . = 0x00006C; gpProcEkrDispUP = .;
@@ -84,11 +84,11 @@
 . = 0x00400C; gUnknown_0200400C = .;
 . = 0x004054; gUnknown_02004054 = .;
 . = 0x00407C; gUnknown_0200407C = .;
-. = 0x004088; gUnknown_02004088 = .;
+. = 0x004088; gBanimPal1 = .;
 . = 0x0040AE; gUnknown_020040AE = .;
 . = 0x0040B2; gUnknown_020040B2 = .;
-. = 0x004128; gUnknown_02004128 = .;
-. = 0x0041C8; gUnknown_020041C8 = .;
+. = 0x004128; gBanimPal2 = .;
+. = 0x0041C8; gBanimOaml = .;
 . = 0x00422C; gUnknown_0200422C = .;
 . = 0x004254; gUnknown_02004254 = .;
 . = 0x0044AC; gUnknown_020044AC = .;
@@ -103,7 +103,7 @@
 . = 0x0078D8; gUnknown_020078D8 = .;
 . = 0x008000; gUnknown_02008000 = .;
 . = 0x0087A0; gUnknown_020087A0 = .;
-. = 0x0099C8; gUnknown_020099C8 = .;
+. = 0x0099C8; gBanimOamr2 = .;
 . = 0x00A2D8; gUnknown_0200A2D8 = .;
 . = 0x00A300; gUnknown_0200A300 = .;
 . = 0x00AF00; gUnknown_0200AF00 = .;
@@ -124,12 +124,12 @@
 . = 0x00F158; gUnknown_0200F158 = .;
 . = 0x00F15C; gUnknown_0200F15C = .;
 . = 0x00F17C; gStringBufferAlt = .;
-. = 0x00F1C8; gUnknown_0200F1C8 = .;
+. = 0x00F1C8; gBanimScrLeft = .;
 . = 0x010004; src/sallycursor.o(ewram_data);
 . = 0x01000C; gUnknown_0201000C = .;
 . = 0x01117C; gPrepMainMenuTexts = .;
 . = 0x0111CC; gPrepUnitPool = .;
-. = 0x011BC8; gUnknown_02011BC8 = .;
+. = 0x011BC8; gBanimScrRight = .;
 . = 0x0121CC; gPrepUnitList = .;
 . = 0x0122D4; gUnknown_020122D4 = .;
 . = 0x012914; gUnknown_02012914 = .;
@@ -256,12 +256,12 @@
 . = 0x01FAC8; gUnknown_0201FAC8 = .;
 . = 0x01FACC; gEkrWindowAppearExist = .;
 . = 0x01FAD0; gEkrNamewinAppearExist = .;
-. = 0x01FAD4; gUnknown_0201FAD4 = .;
+. = 0x01FAD4; gProcEkrBaseAppearExist = .;
 . = 0x01FAD8; gUnknown_0201FAD8 = .;
 . = 0x01FADC; gUnknown_0201FADC = .;
 . = 0x01FB04; gUnknown_0201FB04 = .;
 . = 0x01FB0C; gEkrBgXOffset = .;
-. = 0x01FB18; gUnknown_0201FB18 = .;
+. = 0x01FB18; gpProcEfxAnimeDrv = .;
 . = 0x01FB1C; gUnknown_0201FB1C = .;
 . = 0x01FB24; gUnknown_0201FB24 = .;
 . = 0x01FB28; gUnknown_0201FB28 = .;
@@ -572,8 +572,8 @@
 . = 0x03E120; gEkrDistanceType = .;
 . = 0x03E122; gEkrPairBmLoc = .;
 . = 0x03E12A; gAnimRoundData = .;
-. = 0x03E152; gUnknown_0203E152 = .;
-. = 0x03E156; gUnknown_0203E156 = .;
+. = 0x03E152; gEfxPairHpBufOffset = .;
+. = 0x03E156; gEfxHpBarBuf = .;
 . = 0x03E182; gEkrPairBanimID2 = .;
 . = 0x03E188; gpEkrBattleUnitLeft = .;
 . = 0x03E18C; gpEkrBattleUnitRight = .;
@@ -596,7 +596,7 @@
 . = 0x03E1DC; gUnknown_0203E1DC = .;
 . = 0x03E1E0; gUnknown_0203E1E0 = .;
 . = 0x03E1E4; gBaArenaFlag = .;
-. = 0x03E1E8; gUnknown_0203E1E8 = .;
+. = 0x03E1E8; gpProcefxopCur = .;
 . = 0x03E1EC; gUnknown_0203E1EC = .;
 . = 0x03E1F0; gCurrentMapAnimState = .;
 . = 0x03E1F8; gUnknown_0203E1F8 = .;
